# dist/api.yml
...
openapi: 3.0.0
info:
  title: League Managment System API
  description: API for league managments system
  contact:
    email: yitzhakh@post.bgu.ac.il
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  version: 1.0.0
servers:
- url: https://virtserver.swaggerhub.com/itzik-ariel/ItzikArielWebDevelopProjectAPI/1.0.0
  description: API for league managments system
...
paths:
...
  /register:
    post:
      tags:
      - User
      summary: Register new user to system
      description: |
        Sends data to server to register new user.
        user name, first and last name, password, email, country and url to image.
      parameters:
      - name: newUserInfo
        in: header
        required: false
        style: simple
        explode: false
        schema:
          $ref: '#/components/schemas/User'
      responses:
        "200":
          description: OK. All details sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "204":
          description: User name already in use
        "400":
          description: Bad request
  /login:
    post:
      tags:
      - User
      summary: Sign in user to system
      description: |
        Send log in details to server (user name + password), if data is valid sends back all user data.
      parameters:
      - name: userName
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
        example: itzikfa94
      - name: password
        in: header
        required: false
        style: simple
        explode: false
        schema:
          type: string
        example: 100InSvivot
      responses:
        "200":
          description: OK. All details sent
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        "204":
          description: No such user
        "400":
          description: Bad request
...
components:
  schemas:
    User:
      type: object
      properties:
        id:
          type: string
          example: ItzikFa94
        email:
          type: string
          example: yitzhakh@post.bgu.ac.il
        country:
          type: string
          example: Israel
        password:
          type: string
          example: 100InSvivot
        firstName:
          type: string
          example: Itzik
        lastName:
          type: string
          example: Fadida
        imgURL:
          type: string
          example: https://tinyurl.com/2sm7x6rs
        userType:
          type: string
          enum:
          - regular
          - unionDelegate
...